Complete updates

+++ src/app/_views/speech-page/speech-page.component.ts
@@ -4,6 +4,7 @@ import { filter } from 'rxjs/operators';
 import { SpeechService } from '../../_services/speech.service'
 import { FormBase } from '../../_models/form-base'
 import { Observable } from 'rxjs'
+import { delay } from 'rxjs/operators'
 
 import { BsModalService, BsModalRef } from 'ngx-bootstrap/modal'
 
@@ -17,7 +18,11 @@ export class SpeechPageComponent implements OnInit {
   speechId = 0
   inputs$: Observable<FormBase<any>[]>
   addInputs$: Observable<FormBase<any>[]>
+  searchInputs$: Observable<FormBase<any>[]>
   modalRef: BsModalRef
+  isCollapsed = true
+  filteredItems = []
+  selectedItem = {}
 
   constructor(
     private route: Router,
@@ -35,6 +40,7 @@ export class SpeechPageComponent implements OnInit {
   ngOnInit(): void {
     this.service.getSpeeches().subscribe(s => {
       this.items = s
+      this.filteredItems = s
     })
     this.getFormConfig()
   }
@@ -42,34 +48,78 @@ export class SpeechPageComponent implements OnInit {
   set _speechId(value: number) {
     if (this.speechId !== value) {
       this.speechId = value
+      this.selectedItem = this.items.filter(i => i.id === this.speechId)[0]
+      this.getFormConfig()
     }
   }
 
   getFormConfig () {
-    let speechData = this.items.filter(i => i.id === this.speechId)[0]
     this.addInputs$ = this.service.getAddForm()
-    this.inputs$ = this.service.getUpdateForm(speechData)
+    this.inputs$ = this.service.getUpdateForm(this.selectedItem)
+    this.searchInputs$ = this.service.getSearchForm()
   }
 
   onSubmit (data, type) {
     if (type === 'create') {
       this.service.add(data)
-      this.resetForm()
+      this.addInputs$ = this.service.getAddForm()
     } else if (type === 'update') {
       this.service.update(this.speechId, data)
+      alert('Speech Updated!')
     }
+    this.modalRef.hide()
   }
 
   onDelete () {
-    this.service.delete(this.speechId)
+    let proceed = confirm('Are you sure you want to delete this?')
+    if (proceed) {
+      this.service.delete(this.speechId)
+      this.speechId = 0
+      this.resetForm()
+    }
   }
 
-  setSelectedItem () {
-    this._speechId = this.activatedRoute.snapshot.paramMap.get('id') ? parseInt(this.activatedRoute.snapshot.paramMap.get('id')) : 0
+  onSearch (params) {
+    this.filteredItems = this.items
+    let filtered = []
+    if (!this.isEmptyParams(params)) {
+      this.filteredItems.map(fi => {
+        let toPush = false
+        Object.keys(params).map(key => {
+          if (params[key]) {
+            console.log(key, fi[key], params[key], fi[key].toLowerCase().indexOf(params[key].toLowerCase()))
+            toPush = (fi.hasOwnProperty(key) && fi[key].toLowerCase().indexOf(params[key].toLowerCase()) > -1) ? true : false
+          }
+        })
+        if (toPush) filtered.push(fi)
+      })
+      this.filteredItems = filtered
+      let list = document.getElementById('list')
+      list.scrollIntoView()
+    }
+  }
+
+  onShare (e) {
+    console.log(e.target.email.value)
+    if (Object.keys(this.selectedItem).length) {
+      console.log(`mailto:${e.target.email.value}+?subject=Speech&body=${this.selectedItem.content}`)
+    }
+    return false
+  }
+
+  isEmptyParams (params) {
+    for (let key in params) {
+      if (params[key] !== null && params[key] != '') return false
+    }
+    return true
+  }
+
+  setSelectedItem (id) {
+    this._speechId = id
   }
 
   resetForm () {
-    this.inputs$ = this.service.getAddForm()
+    this.inputs$ = this.service.getUpdateForm({})
   }
 
 }
